Revert "refactor(studio): static tab kind registry and tab id codec (PR 1/9) …"
This reverts commit ffd6d6636ddfcb5956d0c1e4799777cbfaf48980.
